move delegate to utilities
