+++ v4/utilities/src/main/java/exchange/dydx/utilities/utils/ActivityDelegate.kt
@@ -1,4 +1,4 @@
-package exchange.dydx.trading.integration.fcm
+package exchange.dydx.utilities.utils
 
 import android.app.Activity
